+++ app/com/arpnetworking/kairos/service/KairosDbServiceImpl.java
@@ -240,81 +240,6 @@ public final class KairosDbServiceImpl implements KairosDbService {
         return response;
     }
 
-    private static void maybeQueueConsistencyChecks(final MetricsQuery query, final Throwable throwable) {
-        if (throwable != null) {
-            // Something downstream in the calling pipeline should get the error.
-            return;
-        }
-
-        if (!query.getStartTime().isPresent()) {
-            // TODO: log
-            return;
-        }
-
-        if (!query.getEndTime().isPresent()) {
-            // TODO: log
-            return;
-        }
-
-        final Instant startTime = query.getStartTime().get();
-        final Instant endTime = query.getEndTime().get();
-
-        query.getMetrics().stream()
-                .map(Metric::getName)
-                .map(RollupMetric::fromRollupMetricName)
-                .forEach(rollupMetricMaybe ->
-                    rollupMetricMaybe.ifPresent(rollupMetric -> {
-                        // TODO: maybe a for loop would be better
-                        periodStreamForInterval(startTime, endTime, rollupMetric)
-                            .map(periodStartTime -> new ConsistencyChecker.Task.Builder()
-                                    .setSourceMetricName(rollupMetric.getBaseMetricName())
-                                    .setRollupMetricName(rollupMetric.getRollupMetricName())
-                                    .setStartTime(periodStartTime)
-                                    .setTrigger(ConsistencyChecker.Task.Trigger.QUERIED)
-                                    .setPeriod(rollupMetric.getPeriod())
-                                    .build())
-                            // TODO: move everything below this line to another func
-                            .randomFilter()
-                            // TODO: wait for consistency checker to process the message before sending more?
-                            .forEach(task -> consistencyChecker.tell(task, ActorRef.noSender()));
-                })
-            );
-    }
-
-    private static Stream<Instant> periodStreamForInterval(Instant startTime, Instant endTime, RollupMetric rollupMetric) {
-        final PeriodIterator periods = new PeriodIterator(startTime, endTime, rollupMetric.getPeriod());
-        return StreamSupport.stream(
-                Spliterators.spliteratorUnknownSize(
-                        periods,
-                        Spliterator.ORDERED)
-                , false);
-    }
-
-
-    static class PeriodIterator implements Iterator<Instant> {
-        Instant _periodStart;
-        final Instant _end;
-        final RollupPeriod _rollupPeriod;
-
-        public PeriodIterator(final Instant _start, final Instant _end, RollupPeriod _rollupPeriod) {
-            this._periodStart = _start;
-            this._end = _end;
-            this._rollupPeriod = _rollupPeriod;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return _periodStart.isBefore(_end);
-        }
-
-        @Override
-        public Instant next() {
-            final Instant ret = _periodStart;
-            _periodStart = _rollupPeriod.nextPeriodStart(_periodStart);
-            return ret;
-        }
-    }
-
     /* package private */ static MetricsQuery useAvailableRollups(
             final List<String> metricNames,
             final MetricsQuery originalQuery,
