+++ app/com/arpnetworking/rollups/QueryConsistencyTaskCreator.java
@@ -1,12 +1,93 @@
 package com.arpnetworking.rollups;
 
+import akka.actor.ActorRef;
+import com.arpnetworking.kairos.client.models.Metric;
 import com.arpnetworking.kairos.client.models.MetricsQuery;
 
+import java.time.Instant;
+import java.util.Iterator;
+import java.util.Spliterator;
+import java.util.Spliterators;
 import java.util.function.Consumer;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 public class QueryConsistencyTaskCreator implements Consumer<MetricsQuery> {
     @Override
-    public void accept(MetricsQuery metricsQuery) {
+    public void accept(MetricsQuery query) {
+        if (!query.getStartTime().isPresent()) {
+            // TODO: log
+            return;
+        }
+
+        if (!query.getEndTime().isPresent()) {
+            // TODO: log
+            return;
+        }
+
+        final Instant startTime = query.getStartTime().get();
+        final Instant endTime = query.getEndTime().get();
+
+        query.getMetrics().stream()
+                .map(Metric::getName)
+                .map(RollupMetric::fromRollupMetricName)
+                .forEach(rollupMetricMaybe ->
+                        rollupMetricMaybe.ifPresent(rollupMetric -> {
+                            // TODO: maybe a for loop would be better
+                            checkerTasks(startTime, endTime, rollupMetric)
+                                    // TODO: move everything below this line to another func
+                                    .randomFilter()
+                                    // TODO: wait for consistency checker to process the message before sending more?
+                                    .forEach(task -> consistencyChecker.tell(task, ActorRef.noSender()));
+                        })
+                );
 
     }
+
+    private static Stream<ConsistencyChecker.Task> checkerTasks(Instant startTime, Instant endTime, RollupMetric rollupMetric) {
+        return periodStreamForInterval(startTime, endTime, rollupMetric.getPeriod())
+                .map(periodStartTime -> new ConsistencyChecker.Task.Builder()
+                        .setSourceMetricName(rollupMetric.getBaseMetricName())
+                        .setRollupMetricName(rollupMetric.getRollupMetricName())
+                        .setStartTime(periodStartTime)
+                        .setTrigger(ConsistencyChecker.Task.Trigger.QUERIED)
+                        .setPeriod(rollupMetric.getPeriod())
+                        .build());
+    }
+
+    private static Stream<Instant> periodStreamForInterval(Instant startTime, Instant endTime, RollupPeriod period) {
+        final PeriodIterator periods = new PeriodIterator(startTime, endTime, period);
+        return StreamSupport.stream(
+                Spliterators.spliteratorUnknownSize(
+                        periods,
+                        Spliterator.ORDERED)
+                , false);
+    }
+
+
+    static class PeriodIterator implements Iterator<Instant> {
+        Instant _periodStart;
+        final Instant _end;
+        final RollupPeriod _rollupPeriod;
+
+        public PeriodIterator(final Instant _start, final Instant _end, RollupPeriod _rollupPeriod) {
+            this._periodStart = _start;
+            this._end = _end;
+            this._rollupPeriod = _rollupPeriod;
+        }
+
+        @Override
+        public boolean hasNext() {
+            return _periodStart.isBefore(_end);
+        }
+
+        @Override
+        public Instant next() {
+            final Instant ret = _periodStart;
+            _periodStart = _rollupPeriod.nextPeriodStart(_periodStart);
+            return ret;
+        }
+    }
+
+
 }
